Migrate kvstore.c unit tests to new test framework.
