+++ src/server.c
@@ -6908,7 +6908,6 @@ struct serverTest {
     {"sds", sdsTest},
     {"dict", dictTest},
     {"listpack", listpackTest},
-    {"kvstore", kvstoreTest},
 };
 serverTestProc *getTestProcByName(const char *name) {
     int numtests = sizeof(serverTests)/sizeof(struct serverTest);
